refactor(sync): flatten folders
This was threatening to become a large nest.
