+++ crates/pathfinder/src/sync.rs
@@ -1,3 +1,9 @@
-#[cfg(feature = "p2p")]
-#[allow(dead_code)]
+#![cfg(feature = "p2p")]
+#![allow(dead_code, unused)]
+
+mod class_definitions;
+mod headers;
 mod p2p;
+mod receipts;
+mod state_updates;
+mod transactions;
