+++ crates/pathfinder/src/sync/class_definitions.rs
